Add beginning to the WS server
This will add basic requests such as authenticating, changing the
password and getting the full user config.

Also added the ability to send events to the WS clients that are
assigned to an user (succesfully authenticated). This can be used for
example to notify clients on a change of the config. Currently only
added an event on prefix changes.