Refactor and improve switcher
Removed duplicate code to check if stream was really low bitrate.

The switcher now fetches the latest stats each second and checks if the
current `switchType` is the same as the previous. The scene wont be
switched until the threshold of the same type has been reached. The same
type count will be reset when a new type has been detected.

This change should remove the unnecessary scene switches. For example
when the stream keeps disconnecting and connecting for a period or when
the stream is currently in a low bitrate state but hits a good bitrate
for a second, these examples would previously cause the switcher to
instantly react and change the scene.

To avoid waiting to switch from offline to the live/low scene, this
algorithm will prioritize switching to the live/low scene.