Fix state being blocked by start
When starting the stream with transcoding the state would be held for as
long as it takes to run the loop. This caused the state not being able
to be read by other threads, which would cause chat commands or anything
else that depended on the state to be blocked until the stream started.