Fix fuzzy matching in switch scene
Fuzzy matching didn't work as expected. The previous library would
choose the first occurence of the word. For example when there are two
scenes: live and live2, if live2 appeared before live and the user
searched for "live" it would return live2.

This patch uses damerau-levenshtein distance to properly match a scene.